(refac): new job tests refactored & added global describe
